requirements.txt
aiohttp==3.9.1
aioredis==2.0.1
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
braceexpand==0.1.7
click==8.1.7
discord.py==2.3.2
dnspython==2.4.2
frozenlist==1.4.1
humanize==4.9.0
idna==3.6
import-expression==1.1.4
importlib-metadata==7.0.1
jishaku==2.5.2
motor==3.3.2
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
typing_extensions==4.9.0
validators==0.22.0
yarl==1.9.4
zipp==3.17.0
